Created placer context for accessing various data structures used by the VPR placer. Refactored place.cpp based on the moved global structures by moving certain interrelated routines into other (new) files.
