+++ vpr/src/place/placer_globals.cpp
@@ -1,0 +1,8 @@
+/**
+ * @file placer_globals.cpp
+ * @brief Defines the global variable `g_placer_ctx` in placer_globals.h
+ */
+
+#include "placer_globals.h"
+
+PlacerContext g_placer_ctx;
