+++ vpr/src/place/placer_globals.h
@@ -1,0 +1,10 @@
+/**
+ * @file placer_globals.h
+ * @brief Declares the accessor variable for key global
+ *        structs that are used everywhere in VPR placer.
+ */
+
+#pragma once
+#include "placer_context.h"
+
+extern PlacerContext g_placer_ctx;
