+++ vpr/src/place/timing_place.h
@@ -34,22 +34,12 @@
  *      calc_relaxed_criticality() in `timing_util.cpp`.
  */
 
-#ifndef TIMING_PLACE
-#define TIMING_PLACE
-
+#pragma once
 #include "vtr_vec_id_set.h"
 #include "timing_info_fwd.h"
 #include "clustered_netlist_utils.h"
 #include "place_delay_model.h"
 #include "vpr_net_pins_matrix.h"
-
-std::unique_ptr<PlaceDelayModel> alloc_lookups_and_criticalities(t_chan_width_dist chan_width_dist,
-                                                                 const t_placer_opts& place_opts,
-                                                                 const t_router_opts& router_opts,
-                                                                 t_det_routing_arch* det_routing_arch,
-                                                                 std::vector<t_segment_inf>& segment_inf,
-                                                                 const t_direct_inf* directs,
-                                                                 const int num_directs);
 
 /**
  * @brief PlacerCriticalities returns the clustered netlist connection criticalities
@@ -611,5 +601,3 @@ class PlacerTimingCosts {
     ///@brief Number of levels in the binary tree.
     size_t num_levels_ = 0;
 };
-
-#endif
